Make the script callable from anywhere
Currently `torch.sh` can only be called by going to the program directory
and running `./torch.sh` directly. This isn't useful when using other build
scripts that automatically generate flamegraphs. Thus, I propose this as
a solution that allows both local usage as well as PATH-compatible usage.

All three scripts need to be in the same path in order to work, as I've not
implemented link following.